+++ apps/mobile-wallet/src/features/send/modals/TokenAmountModal.tsx
@@ -94,6 +94,7 @@ const TokenAmountModal = memo<TokenAmountModalProps>(({ tokenId, onAmountValidat
   return (
     <BottomModal2
       titleAlign="left"
+      notScrollable
       title={
         <ModalHeader>
           <AssetLogo size={18} assetId={token.id} />
@@ -111,7 +112,6 @@ const TokenAmountModal = memo<TokenAmountModalProps>(({ tokenId, onAmountValidat
             placeholder="0"
             keyboardType="numeric"
             autoComplete="off"
-            autoFocus
             allowFontScaling
             fontSize={fontSize}
             style={{