+++ apps/mobile-wallet/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @alephium/mobile-wallet
 
+## 2.4.5
+
+### Patch Changes
+
+- 87d49d2: Fix scroll in address details modal
+- 5c0a994: Allow adding custom dApp URLs to favorites
+- 58da9c2: Fix height of token amount modal in send flow
+- 064d0c0: Truncate dApp URL in sign tx modal
+
 ## 2.4.4
 
 ### Patch Changes